Model refactored for ModelQuery state
